+++ client/package.json
@@ -53,7 +53,6 @@
     "react-icons": "^4.4.0",
     "react-router-dom": "^6.3.0",
     "react-stately": "^3.16.0",
-    "react-zoom-pan-pinch": "^2.1.3",
     "reconnecting-websocket": "^4.4.0",
     "remark-rehype": "^10.1.0",
     "rng": "^0.2.2",
